feat: Add GraphRag query experiment
The goal of this PR is to embed the `@cdktf/provider-aws` such that
given a CloudFormation Resource, a similarity search can find candidate
terraform Resources.

Upstash Indices were created, but as it turns out, Mastra's GraphRAG
works without this. The createGraphRagTool  can be used to give an
Agentic model the capability to search an index on its own using a
vectorStoreName. [Vector Store Prompts](https://mastra.ai/docs/rag/retrieval#vector-store-prompts)